requirements.txt
flask
flask-cors
mongoengine
python-dotenv
uuid